feat(CTE): Moves CTE logic into it's own object.
