+++ lib/dialects/sqlite/index.js
@@ -2,6 +2,7 @@
 
 const _ = require('lodash');
 const AbstractDialect = require('../abstract');
+const AbstractCTE = require('../abstract/cte');
 const ConnectionManager = require('./connection-manager');
 const Query = require('./query');
 const QueryGenerator = require('./query-generator');
@@ -12,7 +13,7 @@ class SqliteDialect extends AbstractDialect {
     super();
     this.sequelize = sequelize;
     this.connectionManager = new ConnectionManager(this, sequelize);
-    this.QueryGenerator = _.extend({}, QueryGenerator, {
+    this.QueryGenerator = _.extend({}, QueryGenerator, AbstractCTE, {
       options: sequelize.options,
       _dialect: this,
       sequelize
